refactor(ios): extract App Remote connection lifecycle
Follow-up to #54 — deduplicate connect/authorizeAndPlay bootstrap, replace
pendingAuthorizeRemote with a unified connectionAttempt model, and move
lifecycle code into dedicated files without changing public API behavior.
